% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{add.data}
\alias{add.data}
\title{Function to add data to the env.data object}
\usage{
add.data(rs.data, this.file, this.merge, breaks)
}
\arguments{
\item{rs.data}{env.data, the object to have data appended to}

\item{this.file}{the file containing the data to append}

\item{this.merge}{the structure of the data for the merge - i.e. "spatial_temporal" for spatial and temporal,
"spatial" for spatial only or "state_temporal" for statewide data that varies by year (i.e. the BBS data)
temporal only would be theoretically possible, but has not been scripted.
For now, it assumes the spatial is county, and temporal is year. Need to generalize this to spatial and temporal.

#**# do I need the field renames, or does the cleanup.garbage step adequately address these issues?
#**# If it doesn't, can I move these corrections there, and streamline this function?

Modified from wnv_hlpr.R}
}
\description{
Function to add data to the env.data object
}
