% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{add.rf1.inputs}
\alias{add.rf1.inputs}
\title{Add data layers specific for the Random Forest model}
\usage{
add.rf1.inputs(env.data, rf1.inputs, breaks)
}
\arguments{
\item{env.data}{The environmental data object}

\item{files.to.add}{A vector containing the full paths for .csv files to merge into the env.data object}

\item{merge.type.vec}{A vector containing the merge type for each of the files to be added}
}
\description{
Modified from the add.other function in wnv_hlpr.R
}
