% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NullModels.R
\name{apply.always.absent.null}
\alias{apply.always.absent.null}
\title{Always Absent Null}
\usage{
apply.always.absent.null(
  train.data,
  target.year,
  n.draws = 1,
  draw.vec = c("DRAW_0001")
)
}
\arguments{
\item{train.data}{Training data to be used to set up the always absent null.
Must have a location field with locations}

\item{n.draws}{the number of draws per location. n.draws will not change the
outcome, as this is deterministic, but it be used to change the structure of
the matrix and make the output parallel to that of the other functions.}
}
\value{
out.df A data frame containing the predictions for each location for the target year
}
\description{
Make predictions of 0 for every location. Does not really warrant a stand-alone
function, but done to be parallel to the other functions
}
