% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NullModels.R
\name{apply.ar1}
\alias{apply.ar1}
\title{Apply Autoregressive 1 model}
\usage{
apply.ar1(train.data, target.year, n.draws, draw.vec)
}
\arguments{
\item{train.data}{Training data to be used to calculate the mean values.
Must have a location field with locations, and a count field with number of cases.}

\item{target.year}{The year for the prediction. Included in the output data
frame to indicate the target year, but otherwise unused}

\item{n.draws}{The number of probabilistic draws to include in the data matrix.
Retained here only for consistency in formatting with the probabilistic methods.}

\item{draw.vec}{A vector of column names for the probabilistic draws.}
}
\value{
out.df A data frame containing the predictions for each location for the target year
}
\description{
Apply an Autoregressive 1 model assuming a normal distribution
(with negative values reclassified to be 0)
}
