% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NullModels.R
\name{apply.null.models}
\alias{apply.null.models}
\title{Apply Null Models}
\usage{
apply.null.models(
  in.data,
  n.draws,
  nulls = "ALL",
  out.path = NA,
  in.seed = 20210622
)
}
\arguments{
\item{in.data}{A data set containing 'location' field with locations,
annual human cases in a 'count' field, and year in a 'year' field. Note that
field names must match EXACTLY. For Stratified Incidence and Random Incidence,
human population also needs to be included in a 'POP' field.}

\item{n.draws}{The number of probabilistic draws to make for each location.}

\item{nulls}{Which null models to run. Default of "ALL" runs all models.
Individual models can be run with 2-letter abbreviations in a vector:\tabular{ll}{
HN \tab Historical Null\cr
NB \tab Negative Binomial\cr
AA \tab Always Absent\cr
SI \tab Stratified Incidence\cr
RI \tab Random Incidence\cr
AR \tab Autoregressive 1\cr
SM \tab Stratified Mean\cr
PY \tab Prior Year\cr
UN \tab Uniform Null\cr
MV \tab Mean Value\cr }}

\item{out.path}{The path to write the crps.df and time.df data objects.
If no output is desired, set this to NA (default).}

\item{in.seed}{The starting seed to ensure replicability of the random
processes}
}
\value{
A list with four elements: crps.df; time.df; sample.size.vec, and a list
of model specific results. crps df is a data
frame with the average CRPS score for that year for each null model, while
time.df contains the timing for each null model. sample.size.vec contains
a list of sample sizes by year, while the model.lists object contains the
individual county CRPS scores by year.
}
\description{
Runs the null models as a group and create an joint output data frame and a
timing data frame. NOTE: the always absent and mean value were NOT calculated
using the apply.always.absent.null function or the apply.mean.value.null
functions - these were created later and have not been edited into the code.
}
