% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NullModels.R
\name{apply.uniform}
\alias{apply.uniform}
\title{Uniform Null Model}
\usage{
apply.uniform(train.data, target.year, n.draws, draw.vec)
}
\arguments{
\item{train.data}{Training data to be used to generate the null predictions
Must have a location field with locations, and a count field with number of cases.}

\item{target.year}{The year for the prediction. Included in the output data
frame to indicate the target year, but otherwise unused}

\item{n.draws}{The number of probabilistic draws to include in the data matrix.}

\item{draw.vec}{A vector of column names for the probabilistic draws.}
}
\value{
out.df A data frame containing the forecast for each location for the target year
}
\description{
Apply a uniform distribution between 0 and the maximum number of cases observed
stratified by county.
}
