% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{check.dependencies}
\alias{check.dependencies}
\title{Check that required packages are installed prior to running the code}
\usage{
check.dependencies(model.name, packages)
}
\description{
Check that required packages are installed prior to running the code
}
