% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NullModels.R
\name{check.fields}
\alias{check.fields}
\title{Check Fields}
\usage{
check.fields(in.data, nulls)
}
\arguments{
\item{in.data}{An input data set, see apply.null.models documentation for details.}
}
\description{
Check that input fields are correct in the in.data object. Stop execution
with an informative error if not.
}
