% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{check.models}
\alias{check.models}
\title{Check models}
\usage{
check.models(models.to.run, models.in.dfmip)
}
\arguments{
\item{models.to.run}{The list of models to be run}

\item{models.in.dfmip}{The list of models scripted in dfmip, as a vector}
}
\value{
NONE: this function throws an error if something is wrong.
}
\description{
Check that the selected model names are supported by the dfmip code. Main aim is to catch minor typos, such as entering 'NULL' instead of 'NULL.MODELS'.
}
