% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{cleanup.garbage}
\alias{cleanup.garbage}
\title{Cleanup Garbage}
\usage{
cleanup.garbage(rs.data)
}
\description{
Taken from wnv_hlpr.R, rs.data is env.data object, but with the name used in wnv_hlpr.R
}
\details{
Quick function to restore original field names after the merge
}
