% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{convert.env.data}
\alias{convert.env.data}
\title{Convert environmental data to the seasonal format used by the RF1 model}
\usage{
convert.env.data(weather.data, all.counties, season.breaks)
}
\description{
Convert environmental data to the seasonal format used by the RF1 model
}
