% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{convert.human.data}
\alias{convert.human.data}
\title{Convert human data}
\usage{
convert.human.data(hd, all.counties, all.years)
}
\description{
Simple little function to aggregate human case data by year and county
}
