% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{create.month.day.vecs}
\alias{create.month.day.vecs}
\title{Create a vector of days and months, starting with the first Sunday}
\usage{
create.month.day.vecs(year)
}
\arguments{
\item{year}{The input year to evaluate (must be >1900 and <2100)}
}
\value{
A list containing a vector of numeric months and a vector with numeric days of the months
}
\description{
See also \code{\link{get.start.week}}
NOTE: Could generalize, by making the start day an input.
}
\examples{
month.day.vecs = create.month.day.vecs(2000)
month.day.vecs = create.month.day.vecs(2012)
month.day.vecs = create.month.day.vecs(2013)
month.day.vecs = create.month.day.vecs(2014)
month.day.vecs = create.month.day.vecs(2015)
month.day.vecs = create.month.day.vecs(2016)
month.day.vecs = create.month.day.vecs(2020)

}
