% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{dfmip.forecast}
\alias{dfmip.forecast}
\title{Disease Forecast Model Intercomparison Project (dfmip) Forecast}
\usage{
dfmip.forecast(forecast.targets, models.to.run, human.data, mosq.data,
  weather.data, districtshapefile, weekinquestion, week.id, results.path,
  arbo.inputs = "none", observed.inputs = "none",
  population.df = "none", rf1.inputs = "none")
}
\arguments{
\item{forecast.targets}{The quantities for which hindcasts are to be made. Options are: \tabular{ll}{
annual.human.cases \tab Number of human cases\cr
seasonal.mosquito.MLE \tab Mosquito infection rate maximum likelihood estimate averaged over the entire season\cr }}

\item{models.to.run}{A string vector of the models to run. Options are: \tabular{ll}{
NULL.MODELS \tab Forecasts based on statewide incidence\cr
ArboMAP \tab Development version ArboMAP forecasts, see details section.\cr
ArboMAP.MOD \tab Modified version of ArboMAP model.\cr
RF1_C \tab Random Forest model, climate inputs only (i.e. equivalent inputs to ArboMAP).\cr
RF1_A \tab Random Forest model, all available inputs.\cr  }
Note these entries are case-sensitive and are run by keyword, so run in a fixed order (NULL.MODELS, ArboMAP, ArbMAP.MOD, RF1_C, RF1_A),
regardless of the order specified in the models.to.run vector.}

\item{human.data}{Data on human cases of the disease. Must be formatted with two columns: district and date (in format M/D/Y, with forward slashes as delimiters). An optional 'year' column may also be present. If absent, it will be generated from the date column. The district column contains the spatial unit (typically county), while the date corresponds to the date of the onset of symptoms for the human case.
#**# WHAT IF THESE DATA ARE MISSING? I.E. just making a mosquito forecast with RF1?}

\item{mosq.data}{Data on mosquito pools tested for the disease. Must be formatted with 4 columns: district (the spatial unit, e.g. county), col_date: the date the mosquitoes were tested, wnv_result: whether or not the pool was positive, pool_size: the number of mosquitoes tested in the pool. A fifth column species is optional but is not used by the code}

\item{weather.data}{Data on weather variables to be included in the analysis. See the read.weather.data function for details about data format.
The read.weather.data function from ArboMAP is a useful way to process one or more data files downloaded via Google Earth Engine.}

\item{districtshapefile}{The shapefile with polygons representing the districts. #**# Are there required fields?}

\item{weekinquestion}{The focal week for the forecast. For the Random Forest model, this will be the last day used for making the forecast}

\item{week.id}{A two-part ID for the analysis run to distinguish it from other weeks or runs. The first part is a character string, while the second part is a date in the format YYYY-MM-DD #**# analysis.id would be a better name, but would require changes to the code in multiple places}

\item{results.path}{The base path in which to place the modeling results. Some models will create sub-folders for model specific results}

\item{arbo.inputs}{Inputs specific to the ArboMAP model. If the ArboMAP model or ArboMAP.MOD model are not being run, this should be set to 'none' or omitted from the function call}

\item{observed.inputs}{Observed values to be used for comparison purposes. Not relevant for forecasts, but of interest for evaluating hindcasts. If unused, set the variable to NA or omit from the function call}

\item{population.df}{Census information for calculating incidence. Can be set to 'none' or omitted from the function call #**# NEEDS FORMAT INSTRUCTIONS}

\item{rf1.inputs}{Inputs specific to the RF1 model, see \code{\link{rf1.inputs}}. If this model is not included, this should be set to 'none' or omitted from the function call #**# LINK TO AN OBJECT WITH MORE DETAILS}
}
\value{
dfmip.outputs: List of two objects: forecasts.df (see above) and other.results
other.results contains model-specific results that do not fit in the forecasts.df format
and can be accessed by keywords. For example, for the RF1 model, other.results$rf1 will
return a list. The first list entry contains the mosquito results, the second contains the human results.
See Random Forest 1 Model outputs (below) for details
}
\description{
Generate forecasts for multiple mosquito-borne disease models using a single interface.
It is currently configured to give estimates of human cases from the ArboMAP model and a Random Forest model
as implemented by Keyel et al. 2019 (see details for citations)
}
\details{
ArboMAP Details: See \code{\link[ArboMAP]{ArboMAP}} for more details
   ArboMAP uses a distributed lags statistical approach to forecast vector-borne disease risk
   based on historical and current mosquito surveillance, historical and current weather data,
   and historical numbers of human cases.

   The original code for ArboMAP is available: www.github.com/ecograph/arbomap
   The fork compatible with dfmip is available: www.github.com/akeyel/arbomap/ArboMAP_package
   Citations:
   Davis et al. 2017 PLoS Currents Outbreaks 9 10.1371/currents.outbreaks.90e80717c4e67e1a830f17feeaaf85de.
   Davis et al. 2018 Acta Tropica 185: 242-250

RF1 Details: See rf1 package: ?rf1

Null models are at the scale of the state, but future updates aim to allow variation by district
}
