% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\docType{package}
\name{dfmip}
\alias{dfmip}
\title{dfmip: Disease Forecast Model Inter-comparison Project package}
\description{
dfmip runs multiple disease forecasting models for the purpose of
comparing and contrasting methods. Currently designed for mosquito-borne
diseases, the aim is for future versions to be more generally applicable
}
