% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{dfmip.to.cdc.challenge.format}
\alias{dfmip.to.cdc.challenge.format}
\title{Convert to CDC Forecast Challenge}
\usage{
dfmip.to.cdc.challenge.format(
  model.name,
  forecast.distributions,
  out.file,
  n.draws,
  bins = c(0, 1, 6, 11, 16, 21, 26, 31, 36, 41, 46, 51, 101, 151, 201, 1000)
)
}
\arguments{
\item{model.name}{The name of the model results to subset from the forecast.distributions object}

\item{forecast.distributions}{The forecast.distributions output created by dfmip for a single forecast model}

\item{out.file}{The file to contain the CDC WNV forecast}

\item{n.draws}{The number of forecast draws in the forecast.distributions dataframe}

\item{bins}{Identify bins used. Inclusive on the lower bound. The upper bin must exceed all observations, otherwise probabilities will not sum to 1.}
}
\value{
forecast.template The forecast in the format required by the CDC (also written to file).
}
\description{
Convert to CDC Forecast Challenge
}
