% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{do.rf}
\alias{do.rf}
\title{A random forest based disease forecasting approach}
\usage{
\method{do}{rf}(trap.data, dep.var, independent.vars, results.path,
  do.spatial = 1, create.test.set = 1, create.ci.null = 1,
  label = "", response.type = "continuous", exploratory = TRUE,
  input.seed = 20180830, temporal.field = "year",
  spatial.field = "district")
}
\description{
The do.rf function runs the random forest portion of the code
}
\details{
Majority of code was writen in wnv_hlpr.R; January - June 2018.
Transferred to this file & modified beginning in March 2019. Code adapted to
use common inputs with the ArboMAP code of Davis & Wimberly

#**# Add parameter descriptions.
}
