% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{do.variance.partitioning}
\alias{do.variance.partitioning}
\title{Determine the contribution of each variable to the total variation explained}
\usage{
\method{do}{variance.partitioning}(trap.data, dep.var, best.vars, best.m,
  temporal.accuracy.R2, drop.thresholds, correlation.threshold,
  response.type, do.spatial, results.path, spatial.resolution,
  temporal.resolution, this.label, temporal.field)
}
\description{
Determine the contribution of each variable to the total variation explained
}
