% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE_IR.R
\name{dprev}
\alias{dprev}
\title{dprev}
\usage{
dprev(yes = c(0), no = c(0), disp = "y", conf = 0.95)
}
\arguments{
\item{yes}{a vector of positive pools, with the elements indicating how many
mosquitoes were in each positive pool}

\item{no}{a vector of negative pools, with the elements indicating how many
mosquitoes were in each negative pool}

\item{disp}{'y' indicates to print results to dipslay}

\item{conf}{The confidence interval to return. Default is 0.95 to yield a
95% confidence interval}
}
\description{
Calculate mosquito infection rates
}
\details{
Citation: C. Williams and C. Moffitt 2005. Estimation of pathogen
prevalence in pooled samples using maximum likelihood methods and
open source software. Journal of Aquatic Animal Health 17: 386 - 391
}
\examples{
dprev(y = c(15), n = c(15,15,16))
dprev(c(5,10), c(5))

}
