% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{estimate.custom.r2}
\alias{estimate.custom.r2}
\title{Calculate R2 estimate of variation explained}
\usage{
estimate.custom.r2(observed, predicted)
}
\description{
This R2 is not bounded as zero, as the model is not derived from the original data
The model is derived from a novel data set, consquently, it could fail to explain even
as much variation as the data mean. In these cases, negative R2's will be obtained
}
\details{
Consequently, a negative R2 might be better than no knowledge at all, but it's not apt to be very good (depends on how variable the data are from the mean!)
 Basic assumption is that there is some variability around the mean, and that typically, a model should be able to explain that variation better than a static mean.
}
