% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{expand.human.data}
\alias{expand.human.data}
\title{Expand human data}
\usage{
expand.human.data(cases, arbitrary.date = "08-01",
  case.field = "count", year.field = "year",
  location.field = "location")
}
\arguments{
\item{cases}{Human cases summarized by location and year}

\item{arbitrary.date}{The month and day to fill into the returned date field. This will be the same for all entries}

\item{case.field}{The field containing counts of human cases}

\item{year.field}{The field containing the year of the human cases}

\item{location.field}{The field containing the spatial extent of the human cases (e.g., county)}
}
\value{
human.data Data in the standardized dfmip input format, with a column for date, and a column for location.
location-years with no cases are omitted, consequently an assumption of no missing data within years is made.
}
\description{
Takes human cases from a single row for each county and year, and adds individual rows for each county reporting cases
I.e. converts it into the format used by dfmip. This is unfortunate, as ArboMAP is the only model that requires the
detailed format, and this is not suitable for ArboMAP, as the dates are not accurate. However, it enables a common
workflow for all models, and that is highly desirable (i.e. the goal of the project)
}
