% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{FormatDataForRF1}
\alias{FormatDataForRF1}
\title{Convert data in the format for ArboMAP to a format that can be used by the
RF1 model}
\usage{
FormatDataForRF1(human.data, mosq.data, weekinquestion, weather.data,
  rf1.inputs, results.path, break.type)
}
\arguments{
\item{human.data}{human data from the main input, but read into memory}

\item{mosq.data}{mosquito data from main input, but read into memory}

\item{weekinquestion}{#**# ADD}

\item{weather.data}{non-temporal data from main input}

\item{rf1.inputs}{see \code{\link{rf1.inputs}}}

\item{results.path}{#**# ADD}

\item{break.type}{#**# ADD}
}
\description{
Convert data in the format for ArboMAP to a format that can be used by the
RF1 model
}
