% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_accuracy.R
\name{generate.test.data}
\alias{generate.test.data}
\title{Generate test data}
\usage{
generate.test.data(forecast.target, distribution, sample.size)
}
\description{
NOT USED, NOT CALLED. JUNK FUNCTION INITIALLY INTENDED TO HELP WITH SIMULATED DATA AND TESTING
}
