% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{get.best.m}
\alias{get.best.m}
\title{Function to identify the m value that produces the best results}
\usage{
get.best.m(f, independent.vars, trap.data, response.type)
}
\description{
Function to identify the m value that produces the best results
}
