% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{get.days}
\alias{get.days}
\title{Function to get number of days in a year (copied from wnv_hlpr.R)}
\usage{
get.days(year)
}
\arguments{
\item{year}{The year to examine}
}
\value{
The number of days in the year (365 for non-leap years, 366 for a leap year)
}
\description{
Does not work for years 1900 and before or after 2100 (i.e. it does not handle the Century cases)
}
