% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{get.start.week}
\alias{get.start.week}
\title{Goal is to be able to have predictions comparable to ArboMAP}
\usage{
get.start.week(year)
}
\description{
Goal is to be able to have predictions comparable to ArboMAP
}
