% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{get_sampling_weeks}
\alias{get_sampling_weeks}
\title{Get Weeks to sample}
\usage{
get_sampling_weeks(
  year,
  season_start_month,
  weeks_in_season,
  sample_frequency = 1
)
}
\arguments{
\item{year}{The year in question}

\item{season_start_month}{The first month of the mosquito season, as a number. E.g., July would be 7.}

\item{weeks_in_season}{The number of weeks to sample}

\item{sample_frequency}{How frequently sample (default, 1 = weekly) #**# Other options are not currently supported}
}
\value{
A list of three vectors: month.vec has months, day.vec has days, and year.vec contains years. These are all parallel to each other and need to be the same length
}
\description{
Create vectors of month, day, and year to use for running forecasts.
Samples begin on Sunday of the week, starting with the first Sunday of the month
}
