% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects.R
\docType{data}
\name{human.data}
\alias{human.data}
\title{Human Data}
\format{
An object of class \code{data.frame} with 1326 rows and 2 columns.
}
\source{
\url{https://www.github.com/Ecograph/ArboMAP/}
}
\usage{
human.data
}
\description{
Simulated human case data for South Dakota
}
\keyword{datasets}
