% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE_IR.R
\name{llprevr}
\alias{llprevr}
\title{llprevr}
\usage{
llprevr(p, yes = c(0), no = c(0))
}
\description{
Internal function used in calculating infection rate
}
\details{
Citation: C. Williams and C. Moffitt 2005. Estimation of pathogen
prevalence in pooled samples using maximum likelihood methods and
open source software. Journal of Aquatic Animal Health 17: 386 - 391
}
