% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{mean.incidence.model}
\alias{mean.incidence.model}
\title{Mean Incidence Model}
\usage{
\method{mean}{incidence.model}(human.data, population.df)
}
\arguments{
\item{human.data}{Standardized human data input}

\item{population.df}{A vector with districts (counties) and their populations. A column labeled SPATIAL should contain the county/district information, while population should be in a TOTAL_POPULATION column.}
}
\description{
Predict the number of human cases of WNV based on historical number of human cases, adjusted for population
This model does not adjust risk based on county - clearly an incorrect model assumption, but one that is of interest
}
