% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE_IR.R
\name{MLE_IR}
\alias{MLE_IR}
\title{Calculate Maximum Likelihood Estimate of Infection Rate using R}
\description{
Calculate the MLE using the dprev function
}
\details{
Citation: C. Williams and C. Moffitt 2005. Estimation of pathogen
prevalence in pooled samples using maximum likelihood methods and
open source software. Journal of Aquatic Animal Health 17: 386 - 391

Documentation and minor reformatting performed by
A. Keyel <akeyel at albany.edu> Permission to distribute under a GPL 3
license granted on 2019 March 5 by Chris Williams via email to A. Keyel
}
\author{
C. Williams <chrisw at uidaho.edu> and C. Moffitt <cmoffitt at uidaho.edu>
}
