% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects.R
\docType{data}
\name{mosq.data}
\alias{mosq.data}
\title{Mosquito Data}
\format{An object of class \code{data.frame} with 16274 rows and 5 columns.}
\source{
\url{https://www.github.com/Ecograph/ArboMAP/}
}
\usage{
mosq.data
}
\description{
Simulated mosquito data for South Dakota
}
\keyword{datasets}
