% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{null.validation}
\alias{null.validation}
\title{Similar to systematic, but for the production of the null comparisons}
\usage{
null.validation(trap.data, dep.var, drop.field, create.ci.null)
}
\description{
Similar to systematic, but for the production of the null comparisons
}
