% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_accuracy.R
\name{Output_accuracy_assessment}
\alias{Output_accuracy_assessment}
\title{Output accuracy assessment}
\description{
This R file is designed to compile the accuracy assessments for dfmip
in order to provide a standardized format across forecast models
}
