% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{plotbins}
\alias{plotbins}
\title{Bin plotting function for the cdc forecast bins}
\usage{
plotbins(cdc.bins, out.pdf)
}
\arguments{
\item{cdc.bins}{Data in the CDC template bin format}

\item{out.pdf}{The pdf of plots to be generated}
}
\description{
Create a bar plot for each bin to visualize the predictions. Predictions are inclusive of the bottom number,
and exclusive of the start of the next bin
}
