% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{rf1.inputs}
\alias{rf1.inputs}
\title{rf1.inputs}
\arguments{
\item{files.to.add}{A vector of file names of other data sources to be
included in the Random Forest model. If no additional data is to be added,
this should be an empty vector (i.e. c()) or NA.}

\item{merge.type.vec}{A vector identifying how the files should be joined to
the mosquito and human data. Options are spatial_temporal where merges will
be performed on county and year (e.g., climate data); state_temporal where
merges will be performed on state and year (e.g. BBS data at state level),
and spatial, where merges will be performed on county only (e.g., landcover
and census data). This must have the same length as files.to.add, and if no
additional data is to be added, this should be an empty vector (i.e. c()) or NA.}

\item{analysis.counties}{A list of counties included in the analysis (this
is for ensuring that a county and year that does not have a human case is
treated as a 0)}

\item{analysis.years}{A list of years included in the analysis (this is for
ensuring that a county and year that does not have a human case is treated
as a 0 rather than as missing) #**# This is problematic if a specific county
or year is missing.}

\item{user.drop.vars}{A list of independent variables that should be
excluded from the analysis. If no variables should be excluded, this should
be an empty vector c() or NA.}

\item{mosq.model}{The Random Forest model to use for forecasting mosquito
infection rates. If this is to be fitted from the empirical data, this
should be set to NA}

\item{human.model}{The Random Forest model to use for forecasting human
cases. If this is to be fitted from the empirical data, this should be set
to NA.}

\item{analyze.mosquitoes}{Whether (1) or not (0) results should be produced
for mosquitoes. These will only be in the other.results$rf1 object and
will not affect the forecast.df object}

\item{analyze.humans}{Whether (analyze.humans set to 1) or not (set to 0)
results for the human analysis should be produced. If set to 0, the
forecast.df object will return only NA's}

\item{no.data.exceptions}{#**# COMING SOON A list of county-years within the
range of analysis.years and analysis.counties that are missing data rather
than 0's}
}
\description{
The model-specific inputs required to run the RF1 model
}
