% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{rf1.outputs}
\alias{rf1.outputs}
\title{Random Forest 1 Model Outputs}
\description{
other.outputs$rf1 gives a list of 2 objects. The first is the
mosquito results, the second is the human results. Each of these objects are
lists of objects. The first $MODEL is the Random Forest Model object. This
can be used to generate novel predictions. The second $TEMPORAL.ACCURACY is
a summary of the cross-validation accuracy, with $OVERALL giving the overall
accuracy. Other accuracy results include $RMSE, $SPEARMAN, $N, $ERROR.DF
The third is $ SPATIAL.ACCURACY which performs the same cross-validation,
except on spatial units instead of on years. This has the same structure as
$TEMPORAL.ACCURACY. The fourth is $TEMPORAL.NULL, which contains two null
model results for comparison ($NULL.MEAN and $NULL.CI) each with the same
structure as $TEMPORAL.ACCURACY. The fifth is $SPATIAL.NULL, which contains
null model results for spatial structure, following the same structure as
$TEMPORAL.NULL. The sixth object is $RETAINED.VARS, which includes a list of
independent variables that were retained in the final model. (NOTE: These
can also be extracted from the $MODEL object)
}
\details{
#**# Link to documentation from main RF1 description
#**# SWITCH DESCRIPTION TO USE A TABLE/LIST FORMAT
}
