% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{rf1}
\alias{rf1}
\title{Run the rf1 model}
\usage{
rf1(human.data, mosq.data, districtshapefile, weather.data, weekinquestion,
  rf1.inputs, results.path, break.type = "seasonal",
  response.type = "continuous")
}
\arguments{
\item{human.data}{Data on human cases of the disease. Must be formatted with
two columns: district and date. The district column contains the spatial
unit (typically county), while the date corresponds to the date of the
onset of symptoms for the human case.}

\item{mosq.data}{Data on mosquito pools tested for the disease. Must be
formatted with 4 columns: district (the spatial unit, e.g. county),
col_date: the date the mosquitoes were tested, wnv_result: whether or not
the pool was positive, pool_size: the number of mosquitoes tested in the
pool. A fifth column species is optional but is not used by the code}

\item{districtshapefile}{The shapefile with polygons representing the
districts. #**# Is this used by RF1?}

\item{weather.data}{Data on weather variables to be included in the analysis.
See the read.weather.data function for details about data format. The
read.weather.data function from ArboMAP is a useful way to process one or
more data files downloaded via Google Earth Engine.}

\item{weekinquestion}{The focal week for the forecast. For the Random Forest
model, this will be the last day used for making the forecast}

\item{rf1.inputs}{Inputs specific to the RF1 model, see
\code{\link{rf1.inputs}}. If this model is not included, this should be set
to 'none' or omitted from the function call #**# LINK TO AN OBJECT WITH
MORE DETAILS}

\item{results.path}{The base path in which to place the modeling results.
Some models will create sub-folders for model specific results}

\item{break.type}{The temporal frequency to use for the data. The default is
'seasonal' which breaks the environmental data into January, February,
March; April, May, June; July, August, September; October, November,
December. Other options may be supported in the future.}

\item{response.type}{Whether data should be treated as continuous (mosquito
rates, number of cases) or binary (0 or 1).}
}
\description{
The RF1 model first fits a random forest, then excludes all variables with
importance scores below the mean importance (NOTE: This will be problematic
if all input variables are relevant) It then further pares down the variable
list using variance partitioning. It retains only variables that contribute
uniquely to explaining variation in the unmeasured year (via
leave-one-year-out cross-validation). The RF1 model uses the randomForest
package by Liaw & Wiener 2002, that implements the Random Forest method
developed by Breiman 2001. MLE calcualtions from MLE_IR.R were written by
Williams and Moffit 2005 (reformatted by A. Keyel). Note that an arbitrary
starting seed is set, to ensure that results are repeatable.
}
\details{
Citations
   Breiman, L. 2001. Random forests. Machine Learning 45: 5- 32
   Keyel, A.C. et al. 2019 PLOS ONE 14(6): e0217854. https://doi.org/10.1371/journal.pone.0217854
   Liaw & Wiener 2002. Classification and Regression by randomForest. R News 2: 18-22
   Williams, C and C. Moffitt 2005. Estimation of pathogen prevalence in pooled samples
   using maximum likelihood methods and open source software. Journal of Aquatic Animal Health 17: 386-391
}
