% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{run.dfmip}
\alias{run.dfmip}
\title{Run Disease Forecast Model Intercomparison Project (dfmip)}
\usage{
run.dfmip(forecasts.df, models.to.run, human.data, mosq.data, weather.data,
  districtshapefile, weekinquestion, week.id, results.path,
  arbo.inputs = "none", observed.inputs = "none",
  population.df = "none", rf1.inputs = "none")
}
\arguments{
\item{forecasts.df}{A data frame that contains prior model results (e.g. for
a different week, state or model). If there are no prior results, this
should be initialized as NA and it will be generated by dfmip}

\item{models.to.run}{A string vector of the models to run. Options are
NULL.MODELS, ArboMAP, ArboMAP.MOD, RF1. Note these entries are
case-sensitive. NULL.MODELS make forecasts based on statewide incidence. Future versions may support district-specific results as well
ArboMAP returns the original ArboMAP forecasts, see details for citations
ArboMAP.MOD produces modified ArboMAP forecasts for annual and weekly number of human cases
RF1 applies the Random Forest method used in Keyel et al. 2019 PLOS ONE
Note that the models are run by keyword, so run in a fixed order (NULL.MODELS, ArboMAP, ArbMAP.MOD, RF1),
regardless of the order specified in the models.to.run vector.}

\item{human.data}{Data on human cases of the disease. Must be formatted with two columns: district and date. The district column contains the spatial unit (typically county), while the date corresponds to the date of the onset of symptoms for the human case.
#**# WHAT IF THESE DATA ARE MISSING? I.E. just making a mosquito forecast with RF1?
#**# Does the date need to be in a particular format? The sample data is "/" delimited.}

\item{mosq.data}{Data on mosquito pools tested for the disease. Must be formatted with 4 columns: district (the spatial unit, e.g. county), col_date: the date the mosquitoes were tested, wnv_result: whether or not the pool was positive, pool_size: the number of mosquitoes tested in the pool. A fifth column species is optional but is not used by the code}

\item{weather.data}{Data on weather variables to be included in the analysis. See the read.weather.data function for details about data format.
The read.weather.data function from ArboMAP is a useful way to process one or more data files downloaded via Google Earth Engine.}

\item{districtshapefile}{The shapefile with polygons representing the districts. #**# Are there required fields?}

\item{weekinquestion}{The focal week for the forecast. For the Random Forest model, this will be the last day used for making the forecast}

\item{week.id}{An arbitrary ID for the analysis run to distinguish it from other weeks or runs for different locations. #**# analysis.id would be a better name, but would require changes to the code in multiple places}

\item{results.path}{The base path in which to place the modeling results. Some models will create sub-folders for model specific results}

\item{arbo.inputs}{Inputs specific to the ArboMAP model. If the ArboMAP model or ArboMAP.MOD model are not being run, this should be set to 'none' or omitted from the function call}

\item{observed.inputs}{Observed values to be used for comparison purposes. Not relevant for forecasts, but of interest for evaluating hindcasts. If unused, set the variable to NA or omit from the function call}

\item{population.df}{Census information for calculating incidence. Can be set to 'none' or omitted from the function call #**# NEEDS FORMAT INSTRUCTIONS}

\item{rf1.inputs}{Inputs specific to the RF1 model, see \code{\link{rf1.inputs}}. If this model is not included, this should be set to 'none' or omitted from the function call #**# LINK TO AN OBJECT WITH MORE DETAILS}
}
\value{
dfmip.outputs: List of two objects: forecasts.df (see above) and other.results
other.results contains model-specific results that do not fit in the forecasts.df format
and can be accessed by keywords. For example, for the RF1 model, other.results$rf1 will
return a list. The first list entry contains the mosquito results, the second contains the human results.
See Random Forest 1 Model outputs (below) for details
}
\description{
A single function designed to provide a simple, consistent interface to multiple disease forecasting models.
It is currently configured to give estimates of human cases from the ArboMAP model and a Random Forest model
as implemented by Keyel et al. 2019 (see details for citations)
}
\details{
ArboMAP Details: See \code{\link[ArboMAP]{ArboMAP}} for more details
   ArboMAP uses a distributed lags statistical approach to forecast vector-borne disease risk
   based on historical and current mosquito surveillance, historical and current weather data,
   and historical numbers of human cases.

   The original code for ArboMAP is available: www.github.com/ecograph/arbomap
   The fork compatible with dfmip is available: www.github.com/akeyel/arbomap/ArboMAP_package
   Citations:
   Davis et al. 2017 PLoS Currents Outbreaks 9 10.1371/currents.outbreaks.90e80717c4e67e1a830f17feeaaf85de.
   Davis et al. 2018 Acta Tropica 185: 242-250

RF1 Details: See \code{\link{rf1}}
}
