% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{run.partitioning}
\alias{run.partitioning}
\title{Calculate the variance partitioning run}
\usage{
run.partitioning(trap.data, dep.var, remaining.vars, best.m,
  temporal.accuracy.R2, response.type, temporal.field = "TEMPORAL")
}
\description{
Calculate the variance partitioning run
}
