% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{seasonal.incidence}
\alias{seasonal.incidence}
\title{Seasonal Incidence Model}
\usage{
seasonal.incidence(human.data, focal.year, mean.annual.cases)
}
\arguments{
\item{week.start}{The day of year on which the week starts in the focal year
#**# should this just be year? I bet there is a pattern based on year.

Can apply this at the district level by multiplying the estimates by the mean annual cases for each district.
But likely too few in NYS for that to be particularly useful?}
}
\description{
#**# STILL UNDER DEVELOPMENT
#**# NEED TO THINK ABOUT WEEK STARTING POINT
}
