% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{split.text}
\alias{split.text}
\title{Make a function that can be applied with sapply to split text
#**# REPLACE INSTANCES OF THIS WITH splitter function}
\usage{
\method{split}{text}(x, part)
}
\description{
Make a function that can be applied with sapply to split text
#**# REPLACE INSTANCES OF THIS WITH splitter function
}
