% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{splitter}
\alias{splitter}
\title{Split strings using strsplit in a way that can easily be used within sapply}
\usage{
splitter(string, delimiter, position, as.string = 0)
}
\arguments{
\item{string}{The string to be split}

\item{delimiter}{What to use to split the string}

\item{position}{Which part of the string to retain. Only one piece can be retained
To retain multiple pieces, do a separate function call for each piece}

\item{as.string}{0 Result will be numeric, 1 Result will be string}
}
\description{
Split strings using strsplit in a way that can easily be used within sapply
}
