% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{splitter2}
\alias{splitter2}
\title{Split strings using substr in a way that can easily be used within sapply}
\usage{
splitter2(string, start, end)
}
\arguments{
\item{string}{The string to be split}

\item{start}{The starting position for the split}

\item{end}{The ending position for the split}
}
\description{
Split strings using substr in a way that can easily be used within sapply
}
