% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{test.type}
\alias{test.type}
\title{test.type}
\usage{
test.type(vector, test.name)
}
\arguments{
\item{vector}{An input vector to check}

\item{test.name}{A name for the test location in the code}
}
\description{
Simple testing function to identify a bug with rbind in the code. Bug appears to be external to dfmip.R and output_accuracy.R
}
