% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NullModels.R
\name{try.again}
\alias{try.again}
\title{Try to fit a negative binomial, progressively modifying the data to obtain a fit}
\usage{
try.again(loc.subset, location)
}
\arguments{
\item{loc.subset}{the input data set with cases in a 'count' field}

\item{location}{The location of the cases}
}
\value{
fit the fit object from a negative binomial model
}
\description{
Try to fit a negative binomial, progressively modifying the data to obtain a fit
}
