% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{update.df}
\alias{update.df}
\title{Update data frame}
\usage{
\method{update}{df}(forecast.targets, forecasts.df, results.object)
}
\arguments{
\item{forecast.targets}{The targets to include in the output data frame. See \code{\link{dfmip.forecast}} for options.}

\item{forecasts.df}{The data frame object to contain the results from dfmip}

\item{results.object}{The results object to be integrated into forecasts.df}
}
\description{
forecasts.df should be initialized as NA if it does not already exist and contain data.
Otherwise, it is updated with the results. This function provides a standardized format for inputting
data from different models in to a common data frame
}
