% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{update.distribution}
\alias{update.distribution}
\title{Update list of forecast distributions}
\usage{
\method{update}{distribution}(forecast.targets, model.name, forecast.id,
  forecast.distributions, results.object)
}
\arguments{
\item{forecast.targets}{The targets to include in the output data frame. See \code{\link{dfmip.forecast}} for options.}

\item{model.name}{The name of the model run}

\item{forecast.id}{The ID for this forecast. Together with Model Name this should be a unique key}

\item{forecast.distributions}{The list object to contain the forecast distribution information}

\item{results.object}{The results object to be integrated into forecast.distributions}
}
\value{
forecast.distributions A list object containing forecast probability distributions organized by forecast.keys consisting of model names and forecast IDs.
}
\description{
forecasts.df should be initialized as NA if it does not already exist and contain data.
Otherwise, it is updated with the results. This function provides a standardized format for inputting
data from different models in to a common data frame
}
