% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_accuracy.R
\name{update.time.targets}
\alias{update.time.targets}
\title{Update the continuous accuracy targets}
\usage{
\method{update}{time.targets}(accuracy.metrics, predictions.mat,
  observations.vec, threshold)
}
\arguments{
\item{accuracy.metrics}{A list containing accuracy information. The list is
structured \tabular{ll}{
RMSE \tab Root Mean Squared Error\cr
Scaled_RMSE \tab RMSE scaled by the mean observed value\cr
within_percentage \tab A binary accuracy classification, where a forecast within a specified percentage of the observation is considered accurate, otherwise it is inaccurate.\cr
within_threshold \tab A binary accuracy classification, where a forecast within a specified threshold from the observation is considered accurate, otherwise it is inaccurate\cr
within_threshold_or_percentage \tab A binary accuracy classification, where if it is accurate by either the threshold or percentage approaches, it is considered accurate.\cr
AUC \tab Area Under the Curve from the Receiver Operating Characteristic Plot\cr
}}

\item{predictions.mat}{A matrix where each row corresponds to a separate
forecast, and every column corresponds to a forecast realization. In the
case of point forecasts, there will be a single column. In the case of
probablistic forecasts, there will likely be a thousand or more columns.
Note that these are realizations, and not the probabilities associated with
categories.}

\item{observations.vec}{A vector where each entry corresponds to the observed outcome for the corresponding row in predictions.mat.}

\item{threshold}{For continuous and discrete forecasts, a threshold of error to be used in classifying the forecast as "accurate". The default is +/- 1 human case, +/- 1 week, otherwise the default is 0.}
}
\value{
accuracy.metrics An updated accuracy.metrics object
}
\description{
Update the continuous accuracy targets
}
