% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects.R
\docType{data}
\name{weather.data}
\alias{weather.data}
\title{Weather Data}
\format{
An object of class \code{data.frame} with 457908 rows and 11 columns.
}
\source{
\url{https://www.github.com/Ecograph/ArboMAP/}
}
\usage{
weather.data
}
\description{
Data from GridMET for South Dakota
}
\keyword{datasets}
