% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{write.predictions}
\alias{write.predictions}
\title{Write model predictions}
\usage{
write.predictions(trap.data, dep.var, rf.model, results.path,
  spatial.resolution, temporal.resolution, label)
}
\description{
Write model predictions
}
