% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_accuracy.R
\name{assess.accuracy}
\alias{assess.accuracy}
\title{Assess Accuracy}
\usage{
assess.accuracy(predictions.mat, observations.vec, forecast.target,
  threshold = "default", percentage = "default")
}
\arguments{
\item{predictions.mat}{A matrix where each row corresponds to a separate
forecast, and every column corresponds to a forecast realization. In the
case of point forecasts, there will be a single column. In the case of
probabilistic forecasts, there will likely be a thousand or more columns.
Note that these are realizations, and not the probabilities associated with
categories.}

\item{observations.vec}{A vector where each entry corresponds to the observed outcome for the corresponding row in predictions.mat.}

\item{forecast.target}{The quantity being forecast. Forecasts targets are: \itemize{
\item annual.human.cases
\item human_incidence
\item seasonal.mosquito.MLE
\item peak_mosquito_MLE
\item number_positive_pools
\item human_cases_binary
\item positive_pools_binary
\item peak_timing }}

\item{threshold}{For continuous and discrete forecasts, a threshold of error to be used in classifying the forecast as "accurate". The default is +/- 1 human case, +/- 1 week, otherwise the default is 0.}

\item{percentage}{For continuous and discrete forecasts, if the prediction is within the specified percentage of the observed value, the forecast is considered accurate. The default is +/- 25 percent of the observed.}
}
\value{
accuracy.metrics A list containing accuracy information. The list is
  structured \tabular{ll}{
  RMSE \tab Root Mean Squared Error\cr
  Scaled_RMSE \tab RMSE scaled by the mean observed value\cr
  within_percentage \tab A binary accuracy classification, where a forecast within a specified percentage of the observation is considered accurate, otherwise it is inaccurate.\cr
  within_threshold \tab A binary accuracy classification, where a forecast within a specified threshold from the observation is considered accurate, otherwise it is inaccurate\cr
  within_threshold_or_percentage \tab A binary accuracy classification, where if it is accurate by either the threshold or percentage approaches, it is considered accurate.\cr
  AUC \tab Area Under the Curve from the Receiver Operating Characteristic Plot\cr
  }
}
\description{
High level handler function to decide which metrics to compute and then call
the code to compute them. At present, only continuous and discrete outcomes are supported.
}
\examples{
assess.accuracy(matrix(rnorm(1000, 8, 1), ncol = 1), rnorm(1000, 8, 1), "seasonal.mosquito.MLE")

}
