% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{check.dependencies}
\alias{check.dependencies}
\title{Check that required packages are installed prior to running the code}
\usage{
check.dependencies(model.name, packages)
}
\arguments{
\item{model.name}{The name of the model to check dependencies for}

\item{packages}{The packages required for the model in question}
}
\description{
The function will provide a list of any missing packages that need to be installed by the user
}
