% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{date.subset}
\alias{date.subset}
\title{Restrict data to a single date}
\usage{
date.subset(df, end.year, end.month, end.day, date.format)
}
\description{
df must have a column "date" in format YYYY-MM-DD
Note that the date format was slow for string operations, so a datestr field will be added
df must have a column "year" in format YYYY
}
