% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{dfmip.hindcasts}
\alias{dfmip.hindcasts}
\title{DFMIP Hindcasts}
\usage{
dfmip.hindcasts(models.to.run, focal.years, human.data, mosq.data,
  weather.data, districtshapefile, weekinquestion, week.id, results.path,
  arbo.inputs = "none", population.df = "none", rf1.inputs = "none")
}
\description{
Generate hindcasts for a suite of models using common data inputs, and evaluate their accuracy using a standardized suite of validation metrics
}
