% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{get.DOY}
\alias{get.DOY}
\title{Get Day of Year, given year, month, and day
COPIED FROM wnv_hlpr.R}
\usage{
get.DOY(year, month, day)
}
\description{
Get Day of Year, given year, month, and day
COPIED FROM wnv_hlpr.R
}
