% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{get.start.week}
\alias{get.start.week}
\title{Goal is to be able to have predictions comparable to ArboMAP}
\usage{
get.start.week(year)
}
\arguments{
\item{year}{The selected year}
}
\value{
The day of the first Sunday for the selected year
}
\description{
Return the day associated with the first Sunday for the given year
}
