% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects.R, R/dfmip.R
\docType{data}
\name{rf1.inputs}
\alias{rf1.inputs}
\title{Random Forest Inputs}
\format{
An object of class \code{list} of length 9.
}
\usage{
rf1.inputs
}
\description{
Example inputs to facilitate running the random forest example


}
\details{

}
\keyword{datasets}
