% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_accuracy.R
\name{setup.t.p.defaults}
\alias{setup.t.p.defaults}
\title{Set up threshold and percentage defaults}
\usage{
setup.t.p.defaults(forecast.target, threshold, percentage, binary.targets,
  continuous.targets, discrete.targets, time.targets)
}
\description{
Simple function to configure the defaults based on the type of input
}
