% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmip.R
\name{update.df}
\alias{update.df}
\title{Update data frame}
\usage{
\method{update}{df}(forecasts.df, results.object, observed.inputs)
}
\arguments{
\item{forecasts.df}{The data frame object to contain the results from dfmip}

\item{results.object}{The results object to be integrated into forecasts.df}

\item{observed.inputs}{If the model is being used to hindcast, include the
observed values for comparison to the model predictions}
}
\description{
forecasts.df should be initialized as NA if it does not already exist and contain data.
Otherwise, it is updated with the results. This function provides a standardized format for inputting
data from different models in to a common data frame
}
